Finished first pass of drawing functionality
Cards can now be drawn
There is no logic to determine order of deck, and therefore no empty deck logic. these needs to be added
There is no distinguishment, for the hand, as to which player drew the card. should be a simple addition
Need to add another hand for the enemy